Move 'HttpClient' in its own (test) module.
  Now re-use between the e2e and the state-machine specs. Plus, the module is already standalone and well-scoped: focused on the HTTP API.
